refactor auth enums into types file
