--- src/constants/auth.ts
@@ -1,9 +1,0 @@
-enum AUTH_ERROR_CODES {
-  PASSWORD_REQUIRED = "PASSWORD_REQUIRED",
-  PASSWORD_INVALID = "PASSWORD_INVALID",
-  EMAIL_INVALID = "EMAIL_INVALID",
-  OTP_REQUIRED = "OTP_REQUIRED",
-  OTP_INVALID = "OTP_INVALID",
-}
-
-export { AUTH_ERROR_CODES }
